+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Handle non-existent files/directories gracefully
 * Teach scss-lint that opening curly braces should be preceded by one space
 * Teach scss-lint that placeholder names should be lowercase and use hyphens
+* Add `--show-linters` flag for listing all linters available to scss-lint
 
 0.7.1
 * Upgrade Sass gem dependency from 3.2.8 -> 3.2.9